+++ miden-crypto/src/dsa/rpo_falcon512/keys/mod.rs
@@ -4,7 +4,7 @@ use super::{
 };
 
 mod public_key;
-pub use public_key::{PubKeyPoly, PublicKey};
+pub use public_key::PublicKey;
 
 mod secret_key;
 pub use secret_key::SecretKey;